- Changed title of DOT_PRODUCT parameter. - Changed DOT_PRODUCT to DoubleParameter and used default ScoreFormatter configuration.

+++ src/main/java/io/github/mzmine/modules/dataprocessing/id_nist/NistMsSearchTask.java
@@ -127,7 +127,7 @@ public class NistMsSearchTask extends AbstractTask {
   private int progressMax;
 
   // Dot Product cut-offs.
-  private final int minDotProduct;
+  private final Double minDotProduct;
 
   // MS Level.
   private final int msLevel;
